+++ src/main/java/org/codice/imaging/nitf/core/NitfTextSegmentParser.java
@@ -78,7 +78,7 @@ class NitfTextSegmentParser extends AbstractNitfSegmentParser {
                 break;
             case UNKNOWN:
             default:
-                throw new ParseException("Unsupported reader version", reader.getCurrentOffset());
+                throw new ParseException("Unsupported reader version", (int) reader.getCurrentOffset());
         }
     }
